Разделил на модули доработал лабораторную
